moved the weighted point calculations to a seperate stage
