+++ app/src/main/java/dk/scuffed/whiteboardapp/pipeline/stage_combinations/FullCornerDetection.kt
@@ -11,6 +11,7 @@ import dk.scuffed.whiteboardapp.pipeline.stages.bitmap_process_stages.OpenCVLine
 import dk.scuffed.whiteboardapp.pipeline.stages.lines_stages.BiggestSquareStage
 import dk.scuffed.whiteboardapp.pipeline.stages.lines_stages.LinesAngleDiscriminatorStage
 import dk.scuffed.whiteboardapp.pipeline.stages.pipeline_stages.ThreadedBitmapInputPointOutputStage
+import dk.scuffed.whiteboardapp.pipeline.stages.points_stages.WeigthedPointsStage
 import dk.scuffed.whiteboardapp.utils.Vec2Int
 
 /**
@@ -67,6 +68,13 @@ internal fun fullCornerDetection(
             val biggestSquareStage = BiggestSquareStage(
                 horizontalLinesAngleDiscriminatorStage,
                 verticalLinesAngleDiscriminatorStage,
+                pipeline
+            )
+
+            val weightedCornerStage = WeigthedPointsStage(
+                biggestSquareStage,
+                20,
+                5.0f,
                 pipeline
             )
         },
